requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
cryptography==43.0.1
itsdangerous==2.2.0
httpx==0.27.2
redis==5.0.8
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyyaml==6.0.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.141.0
email-validator==2.2.0
structlog==24.4.0
sentry-sdk==2.13.0
